requirements.txt
openai>=1.12.0
python-dotenv>=1.0.0
PyMuPDF>=1.24.0
pytesseract>=0.3.10
Pillow>=10.0.0
langdetect>=1.0.9
streamlit>=1.32.0
fpdf2>=2.7.0
rapidfuzz>=3.6.0
